src/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
langgraph>=0.2.0
langchain>=0.3.0
langsmith>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-google-genai>=2.0.0
chromadb>=0.5.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
